+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Features
 - Share link from another app
+- Preload title of the page
 # 0.1.0 (2020.05.09)
 
 #### Features